Move ColorProvider to different directory
Move the ColorProvider class from controllers/highlight to
providers/colors.
